+++ Shpping/View/ProductList/Componet/ProductListCell.swift
@@ -91,10 +91,10 @@ private extension ProductListCell {
 
 // MARK: Public
 extension ProductListCell {
-    func configCell(_ viewModel: ShopItemsViewModel & DateConvertable) {
+    func configCell(_ viewModel: ShopItemsViewModel) {
         titleLabel.text = viewModel.title
         descriptionLabel.text = viewModel.description
-        createTimeLabel.text = viewModel.toString(by: viewModel.createTime)
+        createTimeLabel.text = viewModel.createTime.toString()
         priceLabel.text = viewModel.price.formatePrice()
         pictureImageView.kf.setImage(with: viewModel.image, placeholder: UIImage(named: "warning"), options: nil, completionHandler: nil)
     }
